backend phase 1 changes

+++ server/controllers/taskControllers.js
@@ -1,24 +1,35 @@
 const asyncHandler = require('express-async-handler');
-
 const Task = require('../models/taskModel');
+const User = require('../models/userModel');
 
-const createTasks = asyncHandler(async (req, res) => {
-  console.log(req.body);
+const createTask = asyncHandler(async (req, res) => {
   if (!req.body) {
     res.status(400);
     throw new Error('Please add required task details.');
   }
   const task = await Task.create({
+    user: req.user.id,
     taskName: req.body.taskName,
     taskDetail: req.body.taskDetail,
     deadline: req.body.deadline,
   });
-  res.status(200).json(task);
+
+  if (task) {
+    res.status(201).json(task);
+  } else {
+    res.status(400);
+    throw new Error('Invalid request.');
+  }
 });
 
 const readTasks = asyncHandler(async (req, res) => {
-  const tasks = await Task.find();
-  res.status(200).json(tasks);
+  const tasks = await Task.findOne({ user: req.user.id });
+  if (tasks) {
+    res.status(201).json(tasks);
+  } else {
+    res.status(400);
+    throw new Error('Invalid request.');
+  }
 });
 
 //Just updating done status
@@ -30,10 +41,25 @@ const updateTasks = asyncHandler(async (req, res) => {
     throw new Error('Task not found');
   }
 
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(400);
+    throw new Error('User not found');
+  }
+  if (user._id != task.user.toString()) {
+    res.status(400);
+    throw new Error('Not authorized.');
+  }
+
   const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
-  res.status(200).json(updatedTask);
+  if (updatedTask) {
+    res.status(201).json(updatedTask);
+  } else {
+    res.status(400);
+    throw new Error('Invalid request.');
+  }
 });
 
 const deleteTasks = asyncHandler(async (req, res) => {
@@ -44,12 +70,27 @@ const deleteTasks = asyncHandler(async (req, res) => {
     throw new Error('Task not found');
   }
 
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(400);
+    throw new Error('User not found');
+  }
+  if (user._id != task.user.toString()) {
+    res.status(400);
+    throw new Error('Not authorized.');
+  }
+
   const deletedTask = await Task.findByIdAndDelete(req.params.id, req.body);
-  res.status(200).json(deletedTask);
+  if (deletedTask) {
+    res.status(201).json(deletedTask);
+  } else {
+    res.status(400);
+    throw new Error('Invalid request.');
+  }
 });
 
 module.exports = {
-  createTasks,
+  createTask,
   readTasks,
   updateTasks,
   deleteTasks,
